+++ src/Blogger/BlogBundle/Resources/public/js/blogs.js
@@ -1,27 +1,92 @@
-function setImageTitle(image)
-{
+function setImageTitle(image) {
     var src = image.attr('src');
     var name = src.substr(src.lastIndexOf('/') + 1);
+    blockUI();
 
     $.ajax({
         type: 'GET',
         url: 'setTitleImage/' + name,
         success: function (data) {
-            image.notify("Изображение установлено на заголовок!", "success");
+            if (data.notice == 'success') {
+                $.unblockUI();
+                image.notify("Изображение установлено на заголовок!", "success");
+            }
         }
+    }).always(function () {
+        $.unblockUI();
     });
 }
 
 function deleteImage(image) {
     var src = image.attr('src');
     var name = src.substr(src.lastIndexOf('/') + 1);
-
+    blockUI();
     $.ajax({
         type: 'GET',
         url: 'deleteImage/' + name,
         success: function (data) {
-            console.log(data.notice);
-            if (data.notice == 'success') location.reload();
+            if (data.notice == 'success') {
+                $.unblockUI();
+                image.remove();
+            }
+        }
+    }).always(function () {
+        $.unblockUI();
+    });
+}
+
+function editBlog(submitForm, files) {
+    var url = $('#update-blog').attr('action');
+    var formData = new FormData();
+    var submitParams = submitForm.serializeArray();
+
+    blockUI();
+
+    $.each(files, function (key, value) {
+        formData.append(key, value);
+    });
+
+    $.each(submitParams, function (key, value) {
+        formData.append(value.name, value.value);
+    });
+
+    $.ajax({
+        type: 'POST',
+        url: url,
+        cache: false,
+        dataType: 'json',
+        processData: false,
+        contentType: false,
+        data: formData,
+        success: function (data) {
+            if (data.notice == 'success') {
+                $.unblockUI();
+                $('#blog-title').text(submitParams[0].value);
+                $('#blog-content').text(submitParams[1].value);
+                $('#blog-tag').text(submitParams[2].value);
+                $('.blog').find('.large').remove();
+                $.each(data.images, function (key, value) {
+                    $('.blog').append('<img class="large"' +
+                        'src="/images/' + value + '"' +
+                        'alt="image not found">');
+                });
+            }
+        }
+    }).always(function () {
+        $.unblockUI();
+    });
+}
+
+function blockUI() {
+    $.blockUI({
+        css: {
+            border: 'none',
+            padding: '15px',
+            backgroundColor: '#000',
+            '-webkit-border-radius': '10px',
+            '-moz-border-radius': '10px',
+            opacity: .5,
+            color: '#fff'
         }
     });
 }
@@ -29,6 +94,7 @@ function deleteImage(image) {
 $(document).ready(function () {
     var setTitleImageMode = false;
     var deleteImageMode = false;
+    var files;
 
     function buttonAction(mode, image, notice) {
         if (mode) {
@@ -38,6 +104,26 @@ $(document).ready(function () {
             image.css('opacity', 1);
             $('.large').removeAttr('title');
         }
+    }
+
+    if ($('.large').length == 0) {
+        $('#but-image').hide();
+        $('#but-image-delete').hide();
+    } else {
+        $('#but-image').show();
+        $('#but-image-delete').show();
+    }
+
+    $("#blogEditType_uploadedFiles").on('change', function () {
+        files = event.target.files;
+    });
+
+    if ($('.large').length == 0) {
+        $('#but-image').hide();
+        $('#but-image-delete').hide();
+    } else {
+        $('#but-image').show();
+        $('#but-image-delete').show();
     }
 
     $('#but-edit').click(function () {
@@ -61,23 +147,28 @@ $(document).ready(function () {
         buttonAction(deleteImageMode, $(this), 'Кликните, чтобы удалить изображение');
     });
 
-    $('.large').mouseover(function () {
+    $('body').on('mouseover', '.large', function () {
         if (setTitleImageMode || deleteImageMode) {
             $(this).attr('id', 'image-action');
         }
     });
 
-    $('.large').mouseout(function () {
+    $('body').on('mouseout', '.large', function () {
         $(this).removeAttr('id');
-    })
+    });
 
-    $('.large').click(function () {
+    $('body').on('click', '.large', function () {
         if (setTitleImageMode) {
-            setImageTitle($(this))
+            setImageTitle($(this));
         }
 
-        if (deleteImageMode) {
+        if (deleteImageMode && confirm('Are you sure?')) {
             deleteImage($(this));
         }
+    });
+
+    $('#update-blog').submit(function () {
+        event.preventDefault();
+        editBlog($(this), files);
     });
 });
